requirements.txt
fastapi
uvicorn
python-telegram-bot
google-genai
tavily-python
supabase
pydantic
pyyaml
httpx
